backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
slowapi
email-validator
